IRGen: give non_abi WeakODR linkage
Mark the non_abi functions as weak_odr.  This is needed on Windows
where the default argument generators are emitted as non-ABI functions.
These can be emitted multiple times and need to be coalesced by the
linker - that is be COMDATed as per WeakODR semantics as multiple
definitions are a hard error with PE/COFF.  Use WeakODR rather than
LinkOnceODR as we want to ensure that the symbol is preserved even if
unreferenced.