[API] Start deriving conformances for `Differentiable`.
Implement `Differentiable` derived conformances for struct types.
This improves the usability of `Differentiable` for user-defined types.

`Differentiable` conformances can be derived for a struct type when all
stored properties conform to `Differentiable`.

Currently, synthesis also requires that the struct conforms to `VectorNumeric`
and that all members have `Self == TangentVector == CotangentVector`.
This limitation will be removed in a follow-up patch.

There are many todos:
- Lift aforementioned limitation and synthesize member `TangentVector`
  and `CotangentVector` structs. This is necessary when not all members
  have `Self == TangentVector == CotangentVector`.
  - Mark synthesized structs with `@_fieldwiseProductSpace`.
  - I'm mostly done implementing this, need to fix some bugs.
- Emit warning when deriving `Diffentiable` conformances for struct type
  that doesn't conform to `VectorNumeric` but could.
  - Warning: "'TangentVector' and 'CotangentVector' could be '<Self>' if
    '<Self>' conforms to 'VectorNumeric'"
- Support `Differentiable` synthesis for enums.