[stdlib] [AutoDiff] Numeric refines VectorNumeric, FloatingPoint refines Differentiable
This is an experimental patch that makes `Numeric` refine `VectorNumeric` and makes `FloatingPoint` refine `Differentiable`.

This patch also declares primitive VJPs for tensor arithmetic operators that take a scalar on one side. Unfortunately, this is blocked on the type checker support for the trailing `where` clause in `@differentiable` attributes ([SR-9596](https://bugs.swift.org/browse/SR-9596)), which should be fixed ASAP.