Parse: explicitly state ordering of parsing (NFC)
When the parameters are inlined, with optimizations, the evaluation
order is under specified (with MSVC, the arguments are evaluated right
to left rather than left to right).  This results in the arguments being
processed incorrectly.  Create local variables to enforce the ordering
of the processed tokens.  Fixes the build on Windows with optimizations.