Sema: Validate extension if needed in isExtensionApplied()
We're looking at the generic signature of the extension, so it must
have been validated at this point.

I don't have a test case for this and maybe it never came up, but
nothing forces it to be validated here so let's make it more robust
by doing it explicitly.