IRGen: prevent WeakExternal on COFF
We cannot use ExternalWeakLinkage on PE/COFF as the semantics for that
do not exist.  It is theoretically possible to emulate this by manually
creating a default value for the symbol that evaluates to 0, but that
requires adding a set of linker directives and variable definitions.
Disable the weak linkage rather than generate an invalid module.