[gyb] Fix Windows path normalization
A number of gyb tests (gyb --test) were failing on Windows. It appears it was
introduced by commit d57b41e which altered the behaviour of line
directives to print paths with forward-slashes even on Windows.

This commit changes two areas of gyb.py:

1. Doctests use normalized paths where applicable.

2. Path normalization is done early in ParseContext initialization
instead of when formatting line directives in ExecutionContext.append_text.

Fixes: SR-9644