[stdlib] [AutoDiff] Temporarily export API `_valueWithPullback(at:_:in:)` for methods.
Differentiating a method should be as simple as calling the method in the trailing closure of a differential operator. However, that is limited due to the fact that AD cannot differentiate generic functions yet. This PR exports differential operator `_valueWithPullback(at:_:in:)`, which wraps `Builtin.autodiffApply_vjp_method`. This makes it possible to define machine learning abstractions that take the derivative of differentiable protocol requirements.