[pmo] Eliminate PartialStore and consider Assign to be something that stops allocations from being removed.
PartialStore and Assign in PMO today both are allowed and do not prevent an
allocation from being removed. This is incorrect with ownership since in both
cases if we have a non-trivial type, we would need to make sure that the value
being overwritten is destroyed.

Clearly, with ownership given that we do not know how to handle assigns in the
aforementioned way, we can not remove allocation with ownership. When I make the
optimization more conservative by changing assigns to stop allocations from
being removed, I see no difference in the tests/benchmarks/etc. So it makes
sense to make the change now.

The reason to remove PartialStore is that:

1. It is actually a vestigal remnant of Definite Init code in Predictable Mem
Opts. If you look in Definite Init, understanding PartialStores is a very
important part of the algorithm for diagnosing partially initialized memory
locations. In Predictable Mem Opts it only controlled whether or not we could
eliminate an allocation.

2. PartialStore assumes that values are always initialized completely so a
partial store must be an assign. Despite that the two places it was used in the
code was a place where we assumed that SILGen was always going to always have an
init anyways (meaning we produced a partial store without need) or in a case
where if we do not have an init, we can just treat it like an assign (the
copy_addr) case.