ClangImporter: make bridging header more portable
`#import` on Windows is problematic at as the default behaviour in C/C++
mode is to be an import of a type library (COM feature).  This can be
adjusted by always enabling Objective-C interop on Windows, but the
runtime is not always built with ObjC interop.

Unfortunately, always using `#include` breaks semantics on Apple where
some places rely on the include-once semantics of `#import`.

For now, control the inclusion technique to be based on whether ObjC
interop is enabled or not.