Added eager mode support to tracer (no longer need to GPE the trace function first).
The design changes are:

1. Added TFE_AddEagerOpToGraph() and some other experimental eager C APIs, to
  help converting an eager op to a trace graph node. This uses a TraceContext to
  keep state -- When an input tensor handle of an eager op is a concrete tensor,
  a symbolic tensor is created for it, representing an input tensor to the trace
  graph function.

e.g. if the function to trace has code:
let x = Tensor(1.0)
let y = x + x

Then x becomes an input tensor to the generated trace graph function, and the
trace context contains a map entry for the concrete tensor handle x and the
corresponding symbolic tensor that wraps the PlaceHolder graph function node to
receive x as an input tensor to the trace graph function.

2. Added a new swift entry point @_swift_tfc_EagerExecute to wrap
  TFE_Execute(). A function call to it is emitted in irgen. In tracing mode, it
  calls TFE_AddEagerOpToGraph() to add the eager op as a trace graph node.

3. In the trace() impl, first create a list of symbolic tensors and feed them
into the function to trace `fn`.